refactor(input_parser): Create submodule
Refactored so that input parsing functions are now in
their own "parser" module.
Extracted most of the functions from "core".
Separated tests accordingly.
